Made Checkered work on Windows.  Removed Windows warnings in mrvImageBrowser and mrvImageInformation and mrvPreferences.
Former-commit-id: 9de9855ce83df444c5de6707d0055d1f96b8de16
Former-commit-id: 30f31382b4623d5ef06b12af8b7272c972bdd011 [formerly 1083bb51c7432f9c96e3211cf6dfe6c4cfd6e96a] [formerly e9a270e5eca68529219cd8d718dd64afb0481ada [formerly ad0f8ea66e32fa056e900cb4d19cea7ec7718ba3]] [formerly feb7344f3ae86643259c30d3ffc141294ee85054] [formerly feb7344f3ae86643259c30d3ffc141294ee85054 [formerly feb7344f3ae86643259c30d3ffc141294ee85054 [formerly 621133f3bd2cff69e715e84f93a8a91e856e3978]]] [formerly a0588999e3546442d55555f1fe6627386949f221 [formerly a0588999e3546442d55555f1fe6627386949f221 [formerly cf96cf7cbbec5be8373b53f801ae65c2466ce240]] [formerly ba8129b42fa8ead6f6c356e48ceef5b949aad5fe] [formerly e5176bb950da3d1668284815488d1dd3cde77279 [formerly ba8129b42fa8ead6f6c356e48ceef5b949aad5fe [formerly ba8129b42fa8ead6f6c356e48ceef5b949aad5fe [formerly 9539944a80f2b4a4c469717139f83fbf5bd07cdd]]]]]
Former-commit-id: 7c750d1cbcf6a47dcd7accc2b171e358e6d9430f [formerly 7c750d1cbcf6a47dcd7accc2b171e358e6d9430f [formerly ab4e68985b2323f7fc7e60c549471cca08de9ea5]] [formerly 185ab8236849a50a171182548012217562d64d67] [formerly 185ab8236849a50a171182548012217562d64d67 [formerly 185ab8236849a50a171182548012217562d64d67 [formerly 90800d2109be57aa83953ca441b0647e202f600b]]]
Former-commit-id: 37e2f4fded28adde9f0051ad968be90f5c88651f [formerly 37e2f4fded28adde9f0051ad968be90f5c88651f [formerly 1e66111d8a95ad6f6de300ac2369bd492455a253]]
Former-commit-id: 37e2f4fded28adde9f0051ad968be90f5c88651f
Former-commit-id: 3ab0881eee3876c1169dbe27ce34ffe5e2d8ef6c [formerly 057e402d1899442a70542e3cdd89727f4aa93358]
Former-commit-id: 19d216a7f84cfcc33d73fb88c3243a00d11cfddd

+++ mrViewer/src/gui/mrvImageBrowser.cpp
@@ -998,7 +998,7 @@ void ImageBrowser::send_reel( const mrv::Reel& reel )
     if ( view() ) view()->send_network( buf );
 }
 
-void ImageBrowser::send_current_image( int idx, const mrv::media& m )
+void ImageBrowser::send_current_image( size_t idx, const mrv::media& m )
 {
     if (!m) return;
 
@@ -1058,7 +1058,7 @@ void ImageBrowser::send_images( const mrv::Reel& reel)
     mrv::MediaList::const_iterator i = reel->images.begin();
     mrv::MediaList::const_iterator e = reel->images.end();
 
-    int idx = 0;
+    size_t idx = 0;
     for ( ; i != e; ++i, ++idx )
     {
         send_current_image( idx, *i );
